# API/swagger.yaml
openapi: 3.0.0
info:
  title: FintechBankApp API
  description: API completa para o sistema bancário FintechBankApp
  version: 1.0.0
  contact:
    name: FintechBankApp Team
    email: support@fintechbankapp.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Servidor de desenvolvimento
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Erro na operação"
...
    Subscription:
      type: object
      description: |
        Assinatura de cobrança recorrente — linha crua da tabela subscriptions
        (sem normalize, ao contrário do RecurringBill): chaves em snake_case e
        amount como string numérica (coluna NUMERIC retornada pelo driver pg).
      properties:
        id:
          type: string
          example: "uuid-xxxx"
        cpf:
          type: string
          example: "11111111111"
        name:
          type: string
          example: "Netflix"
        amount:
          type: string
          description: 'Valor em string numérica (coluna NUMERIC do pg); ex.: "29.90"'
          example: "29.90"
        frequency:
          type: string
          enum: [monthly, yearly]
          description: 'Ciclo de cobrança (computeNextBillingDate: +1 mês ou +1 ano)'
          example: "monthly"
        payment_method:
          type: string
          enum: [debit, credit]
          description: debit → debita users.balance; credit → consome o limite do cartão
          example: "debit"
        status:
          type: string
          enum: [active, cancelled]
          description: active (cobrança recorrente ativa) | cancelled (cancelada pelo cliente)
          example: "active"
        next_billing_date:
          type: string
          format: date-time
          description: Próxima cobrança agendada (consumida pelo cron de assinaturas)
        last_billing_date:
          type: string
          format: date-time
          nullable: true
          description: Última cobrança efetivada (null antes da primeira)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SubscriptionRequest:
      type: object
      required: [name, amount, frequency, payment_method, pin]
      properties:
        name:
          type: string
          example: "Netflix"
        amount:
          type: number
          format: float
          minimum: 0.01
          example: 55.90
        frequency:
          type: string
          enum: [monthly, yearly]
          example: "monthly"
        payment_method:
          type: string
          enum: [debit, credit]
          example: "debit"
        pin:
          type: string
          description: PIN de 4 dígitos (autorização financeira)
          example: "9898"
...
    TransactionCancelRequest:
      type: object
      required: [pin]
      properties:
        pin:
          type: string
          description: PIN de 4 dígitos (autorização financeira)
          example: "9898"
...
    CreditVoucher:
      type: object
      properties:
        id:
          type: string
          example: "uuid-xxxx"
        cpf:
          type: string
          example: "11111111111"
        amount:
          type: number
          format: float
          example: 120.00
        status:
          type: string
          enum: [active, used, expired]
          example: "active"
        source_transaction_id:
          type: string
        source_invoice_id:
          type: string
          nullable: true
        description:
          type: string
        created_at:
          type: string
          format: date-time
        used_at:
          type: string
          format: date-time
          nullable: true
paths:
...
  /subscriptions/{cpf}:
    get:
      tags: [Financeiro]
      summary: Listar assinaturas (cobrança recorrente)
      security:
        - bearerAuth: []
      parameters:
        - name: cpf
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9]{11}$' }
      responses:
        '200':
          description: Lista de assinaturas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  subscriptions:
                    type: array
                    items: { $ref: '#/components/schemas/Subscription' }
        '403':
          description: Acesso negado (CPF de outro usuário sem ser admin)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      tags: [Financeiro]
      summary: Criar assinatura (cobrança recorrente mensal/anual, débito/crédito)
      description: |
        Requer PIN (autorização financeira). Valida bloqueio de cartão/conta
        inadimplente quando payment_method=credit.
      security:
        - bearerAuth: []
      parameters:
        - name: cpf
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9]{11}$' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubscriptionRequest' }
      responses:
        '200':
          description: Assinatura criada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  subscription: { $ref: '#/components/schemas/Subscription' }
        '400':
          description: Payload inválido (frequência/forma de pagamento fora da whitelist, valor <= 0)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Acesso negado, ou cartão bloqueado/conta inadimplente
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /subscriptions/{cpf}/{id}:
    delete:
      tags: [Financeiro]
      summary: Cancelar assinatura (estorna a última cobrança já feita)
      description: |
        Além de parar as cobranças futuras, estorna a última cobrança já feita
        para essa assinatura: débito volta ao saldo; crédito volta como crédito
        direto na fatura se ainda aberta, ou vira credit voucher se a fatura de
        origem já fechou. Se a assinatura nunca foi cobrada, cancela sem
        estorno. Requer PIN.
      security:
        - bearerAuth: []
      parameters:
        - name: cpf
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9]{11}$' }
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionCancelRequest' }
      responses:
        '200':
          description: Assinatura cancelada (com estorno, se havia cobrança anterior)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: "Assinatura cancelada." }
                  reversal:
                    type: object
                    nullable: true
                    properties:
                      id: { type: string }
                      kind: { type: string, enum: [invoice_credit, voucher, debit_refund] }
                      amount: { type: number, format: float }
                      description: { type: string }
                  voucher:
                    $ref: '#/components/schemas/CreditVoucher'
        '404':
          description: Assinatura não encontrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Acesso negado (assinatura não pertence ao CPF)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...